+++ app/gui/src/project-view/stores/graph/index.ts
@@ -25,7 +25,7 @@ import { partition } from '@/util/data/array'
 import { Rect } from '@/util/data/rect'
 import { Err, Ok, mapOk, unwrap, type Result } from '@/util/data/result'
 import { Vec2 } from '@/util/data/vec2'
-import { normalizeQualifiedName, tryQualifiedName } from '@/util/qualifiedName'
+import { normalizeQualifiedName, qnLastSegment, tryQualifiedName } from '@/util/qualifiedName'
 import { useWatchContext } from '@/util/reactivity'
 import { computedAsync } from '@vueuse/core'
 import { map, set } from 'lib0'
@@ -734,6 +734,22 @@ export const { injectFn: useGraphStore, provideFn: provideGraphStore } = createC
       return isAstId(portId) && db.connections.reverseLookup(portId).size > 0
     }
 
+    function nodeCanBeEntered(id: NodeId): boolean {
+      if (!proj.modulePath?.ok) return false
+
+      const expressionInfo = db.getExpressionInfo(id)
+      if (expressionInfo?.methodCall == null) return false
+
+      const definedOnType = tryQualifiedName(expressionInfo.methodCall.methodPointer.definedOnType)
+      const openModuleName = qnLastSegment(proj.modulePath.value)
+      if (definedOnType.ok && qnLastSegment(definedOnType.value) !== openModuleName) {
+        // Cannot enter node that is not defined on current module.
+        // TODO: Support entering nodes in other modules within the same project.
+        return false
+      }
+      return true
+    }
+
     const modulePath: Ref<LsPath | undefined> = computedAsync(
       async () => {
         const rootId = await proj.projectRootId
@@ -789,6 +805,7 @@ export const { injectFn: useGraphStore, provideFn: provideGraphStore } = createC
       addMissingImports,
       addMissingImportsDisregardConflicts,
       isConnectedTarget,
+      nodeCanBeEntered,
       currentMethodPointer() {
         const currentMethod = proj.executionContext.getStackTop()
         if (currentMethod.type === 'ExplicitCall') return currentMethod.methodPointer